meu_projeto/requirements_render.txt
Django==5.2.5
gunicorn==23.0.0
whitenoise==6.10.0
python-dotenv==1.0.0
mercadopago==2.3.0
qrcode==8.2
pillow==10.4.0
cryptography==43.0.3
